Don't explicitly pin multiple constraint dependencies
Before trying to generate a lockfile with an upgraded dependency, Dependabot first explicitly updates the pyproject.toml file like the following:
* It pins the dependency to be upgraded to the resolved versions.
* It pins all other dependencies to the locked version.

In this particular case, the numpy dependency is getting locked to
1.21.6, but this is incorrect, because it's a multiple constraint
dependency, so locking it explicitly will cause some versions to be
removed from the lock file.

In general, the second step does not seem necessary to me. The command
we run to update the lock file, `poetry update <dependency> --lock`
should deal with it fine and keep the locked dependencies not
being upgraded untouched.

However, when removing all that logic, some specs fail since they let
other dependencies be upgraded and I'm not sure why. So as an
alternative solution for the time being, I'm just explicitly avoiding
pinning multiple constraint dependencies.