Fix github_actions pinned to specific refs being downgraded
After fixing the previous crash, I noticed that we were actually cloning
the _repository_ being updated (the one that includes the workflow
files), not the repository of the dependency being updated (the
repository of the action itself).

I have no idea how this somewhat seemed to work in some cases before.