Fix another issue where only default branch commits would be upgraded
The previous logic did not really work with shallow forks. In other for
all branches to be reachable, we need to clone the whole repository.
Also, unless we actually checkout the branches, we need to use `git
branch --remote --contains`.

There's still further work that could be done here, for example, to
disambiguate multiple branches including the commit by checking whether
one of them includes the others, and upgrade to the most recent in that
case. But that's for another time.